+++ lib/flutter_smart_dialog.dart
@@ -10,3 +10,4 @@ export 'src/config/smart_config_toast.dart';
 /// dialog
 export 'src/init_dialog.dart';
 export 'src/smart_dialog.dart';
+export 'src/widget/dialog_scope.dart' show SmartDialogController;